Create S3 Example 1
I ended up with two numerical tables, one for Males, and another for Females The information indicates the number of people that had a certain hair color and what their eye color is. 
I was not able to determine explicitly whether this was and S3 or S4 object using the otype.
But the object was a double. I think this is indicative that it is an S3 object type? also because of the way it was indexed with the brackets. ? [[]]